Feature/clean refactor (#58)
* remove a bunch of v1 things and see if everything still works.

* add refactor_clean to workflows

* typo

* typo 2x

* remove dependencies on v1 that werne't used

* move scripts out of v2 folder

* remove more instances of v2

* log info about which image is changing

* change git helper to look at branch name instead of main for changed files

* Have branch compare itself to just its last commit

* properly use git args for last commit

* coerce change in all notebooks

* 2x ditto

* test sync

* another test

* scipy notebook only build

* remove weird deleted files

* force full rebuild on main

* remove .DS_Store

* remove hardcoded git ref

* change tag to branch name instead of hash

* force rebuild

* replace / with -s in branch name

* test full rebuild

* guard rails for git_ref being named stable

* test full rebuild

* test full rebuild again

* test full rebuild with hash ref

* change git helper and readme
