+++ .github/workflows/main.yml
@@ -6,7 +6,7 @@ env:
 
 on:
   push:
-    branches: [ main, refractor_buildtest]
+    branches: [ main, feature/clean_refactor ]
     paths:
       - "images/**"
       - "model/**"
@@ -96,7 +96,7 @@ jobs:
           GRPC_CERT: ${{ secrets.GRPC_CERT }}
           PREPUSH_IMAGES: '["ucsdets/scipy-ml-notebook"]'
         run: |
-          python3 scripts/v2/main.py
+          python3 scripts/main.py
 
       - name: Push Wiki to GitHub
         if: >
