+++ images/datahub-base-notebook/Dockerfile
@@ -1,6 +1,6 @@
 ARG PYTHON_VERSION=python-3.9.5
 FROM jupyter/datascience-notebook:$PYTHON_VERSION
-# coerce change in all notebooks
+# coerce change in all notebook
 USER root
 
 # see https://github.com/phusion/baseimage-docker/issues/319#issuecomment-1058835363
